refactor: bilibili-share-to-friends 拆分分享弹窗UI组件
