+++ scripts/bilibili-share-to-friends/src/components/Dialog/style.css
@@ -1,0 +1,56 @@
+.bili-share-to-friends-dialog::backdrop {
+  background: rgba(0, 0, 0, 0.38);
+}
+
+.bili-share-to-friends-dialog:not([open]) {
+  display: none !important;
+  pointer-events: none !important;
+}
+
+.bili-share-to-friends-dialog {
+  width: min(420px, calc(100vw - 32px));
+  max-height: min(680px, calc(100vh - 48px));
+  padding: 0;
+  border: 0;
+  border-radius: 8px;
+  background: #fff;
+  color: #18191c;
+  box-shadow: 0 12px 40px rgba(0, 0, 0, 0.18);
+  overflow: hidden;
+  font:
+    14px/1.5 -apple-system,
+    BlinkMacSystemFont,
+    "Segoe UI",
+    "Microsoft YaHei",
+    sans-serif;
+}
+
+.bili-share-to-friends-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 14px 16px;
+  border-bottom: 1px solid #e3e5e7;
+}
+
+.bili-share-to-friends-title {
+  margin: 0;
+  font-size: 16px;
+  font-weight: 600;
+}
+
+.bili-share-to-friends-close {
+  width: 30px;
+  height: 30px;
+  border: 0;
+  border-radius: 4px;
+  background: transparent;
+  color: #61666d;
+  font-size: 22px;
+  line-height: 28px;
+  cursor: pointer;
+}
+
+.bili-share-to-friends-close:hover {
+  background: #f1f2f3;
+}
