+++ scripts/bilibili-share-to-friends/src/components/EntryButton/style.css
@@ -1,0 +1,53 @@
+.bili-share-to-friends-entry {
+  display: inline-flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  width: 60px;
+  min-width: 60px;
+  height: 60px;
+  margin: 0 0px;
+  padding: 0;
+  border: 0;
+  border-radius: 0;
+  color: #18191c;
+  background: transparent;
+  font-size: 12px;
+  line-height: 21px;
+  cursor: pointer;
+  user-select: none;
+  vertical-align: top;
+  transition:
+    color 0.2s,
+    transform 0.2s;
+}
+
+.bili-share-to-friends-entry:hover {
+  color: #00aeec;
+  transform: translateY(-1px);
+}
+
+.bili-share-to-friends-entry-text {
+  color: #18191c;
+  white-space: nowrap;
+}
+
+.bili-share-to-friends-entry:hover .bili-share-to-friends-entry-text {
+  color: #00aeec;
+}
+
+.bili-share-to-friends-entry-icon {
+  width: 36px;
+  height: 36px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 50%;
+  color: #fff;
+  background: #00aeec;
+}
+
+.bili-share-to-friends-entry-icon svg {
+  width: 22px;
+  height: 22px;
+}
