+++ scripts/bilibili-share-to-friends/src/components/StateView/style.css
@@ -1,0 +1,9 @@
+.bili-share-to-friends-state {
+  padding: 42px 18px;
+  color: #61666d;
+  text-align: center;
+}
+
+.bili-share-to-friends-state-error {
+  color: #d03050;
+}
