+++ scripts/bilibili-share-to-friends/src/components/UserList/style.css
@@ -1,0 +1,5 @@
+.bili-share-to-friends-list {
+  margin: 0;
+  padding: 8px 0;
+  list-style: none;
+}
