+++ scripts/bilibili-share-to-friends/src/components/VideoPreview/style.css
@@ -1,0 +1,44 @@
+.bili-share-to-friends-video {
+  display: grid;
+  grid-template-columns: 88px 1fr;
+  gap: 10px;
+  padding: 12px 16px;
+  border-bottom: 1px solid #e3e5e7;
+  background: #f6f7f8;
+}
+
+.bili-share-to-friends-cover {
+  width: 88px;
+  aspect-ratio: 16 / 10;
+  border-radius: 4px;
+  background: #e3e5e7;
+}
+
+.bili-share-to-friends-cover-placeholder {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #9499a0;
+  font-size: 12px;
+}
+
+.bili-share-to-friends-cover-img {
+  object-fit: cover;
+}
+
+.bili-share-to-friends-video-title {
+  margin: 0;
+  color: #18191c;
+  font-size: 13px;
+  line-height: 18px;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+}
+
+.bili-share-to-friends-video-author {
+  margin-top: 6px;
+  color: #9499a0;
+  font-size: 12px;
+}
