+++ scripts/bilibili-share-to-friends/src/styles.css
@@ -1,272 +1,5 @@
-.bili-share-to-friends-entry {
-  display: inline-flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  width: 60px;
-  min-width: 60px;
-  height: 60px;
-  margin: 0 0px;
-  padding: 0;
-  border: 0;
-  border-radius: 0;
-  color: #18191c;
-  background: transparent;
-  font-size: 12px;
-  line-height: 21px;
-  cursor: pointer;
-  user-select: none;
-  vertical-align: top;
-  transition: color .2s, transform .2s;
-}
-
-.bili-share-to-friends-entry:hover {
-  color: #00aeec;
-  transform: translateY(-1px);
-}
-
-.bili-share-to-friends-entry-text {
-  color: #18191c;
-  white-space: nowrap;
-}
-
-.bili-share-to-friends-entry:hover .bili-share-to-friends-entry-text {
-  color: #00aeec;
-}
-
-.bili-share-to-friends-entry-icon {
-  width: 36px;
-  height: 36px;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%;
-  color: #fff;
-  background: #00aeec;
-}
-
-.bili-share-to-friends-entry-icon svg {
-  width: 22px;
-  height: 22px;
-}
-
-.bili-share-to-friends-dialog::backdrop {
-  background: rgba(0, 0, 0, .38);
-}
-
-.bili-share-to-friends-dialog:not([open]) {
-  display: none !important;
-  pointer-events: none !important;
-}
-
-.bili-share-to-friends-dialog {
-  width: min(420px, calc(100vw - 32px));
-  max-height: min(680px, calc(100vh - 48px));
-  padding: 0;
-  border: 0;
-  border-radius: 8px;
-  background: #fff;
-  color: #18191c;
-  box-shadow: 0 12px 40px rgba(0, 0, 0, .18);
-  overflow: hidden;
-  font: 14px/1.5 -apple-system, BlinkMacSystemFont, "Segoe UI", "Microsoft YaHei", sans-serif;
-}
-
-.bili-share-to-friends-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 14px 16px;
-  border-bottom: 1px solid #e3e5e7;
-}
-
-.bili-share-to-friends-title {
-  margin: 0;
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.bili-share-to-friends-close {
-  width: 30px;
-  height: 30px;
-  border: 0;
-  border-radius: 4px;
-  background: transparent;
-  color: #61666d;
-  font-size: 22px;
-  line-height: 28px;
-  cursor: pointer;
-}
-
-.bili-share-to-friends-close:hover {
-  background: #f1f2f3;
-}
-
-.bili-share-to-friends-video {
-  display: grid;
-  grid-template-columns: 88px 1fr;
-  gap: 10px;
-  padding: 12px 16px;
-  border-bottom: 1px solid #e3e5e7;
-  background: #f6f7f8;
-}
-
-.bili-share-to-friends-cover {
-  width: 88px;
-  aspect-ratio: 16 / 10;
-  border-radius: 4px;
-  background: #e3e5e7;
-}
-
-.bili-share-to-friends-cover-placeholder {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: #9499a0;
-  font-size: 12px;
-}
-
-.bili-share-to-friends-cover-img {
-  object-fit: cover;
-}
-
-.bili-share-to-friends-video-title {
-  margin: 0;
-  color: #18191c;
-  font-size: 13px;
-  line-height: 18px;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.bili-share-to-friends-video-author {
-  margin-top: 6px;
-  color: #9499a0;
-  font-size: 12px;
-}
-
 .bili-share-to-friends-body {
   min-height: 180px;
   max-height: 360px;
   overflow: auto;
-}
-
-.bili-share-to-friends-state {
-  padding: 42px 18px;
-  color: #61666d;
-  text-align: center;
-}
-
-.bili-share-to-friends-state-error {
-  color: #d03050;
-}
-
-.bili-share-to-friends-list {
-  margin: 0;
-  padding: 8px 0;
-  list-style: none;
-}
-
-.bili-share-to-friends-person {
-  display: grid;
-  grid-template-columns: 40px 1fr auto;
-  align-items: center;
-  gap: 10px;
-  width: 100%;
-  padding: 10px 16px;
-  border: 0;
-  background: #fff;
-  color: inherit;
-  text-align: left;
-  cursor: pointer;
-}
-
-.bili-share-to-friends-person:hover,
-.bili-share-to-friends-person[aria-selected="true"] {
-  background: #f6fbff;
-}
-
-.bili-share-to-friends-avatar {
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  object-fit: cover;
-  background: #e3e5e7;
-}
-
-.bili-share-to-friends-name {
-  min-width: 0;
-  font-size: 14px;
-  color: #18191c;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.bili-share-to-friends-meta {
-  min-width: 0;
-  margin-top: 2px;
-  color: #9499a0;
-  font-size: 12px;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.bili-share-to-friends-check {
-  width: 18px;
-  height: 18px;
-  border: 1px solid #c9ccd0;
-  border-radius: 50%;
-}
-
-.bili-share-to-friends-person[aria-selected="true"] .bili-share-to-friends-check {
-  border-color: #00aeec;
-  background: radial-gradient(circle at center, #00aeec 0 45%, transparent 47%);
-}
-
-.bili-share-to-friends-footer {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: 10px;
-  padding: 12px 16px;
-  border-top: 1px solid #e3e5e7;
-}
-
-.bili-share-to-friends-btn {
-  min-width: 76px;
-  height: 32px;
-  padding: 0 14px;
-  border: 1px solid #c9ccd0;
-  border-radius: 4px;
-  background: #fff;
-  color: #18191c;
-  cursor: pointer;
-}
-
-.bili-share-to-friends-btn:hover {
-  border-color: #00aeec;
-  color: #00aeec;
-}
-
-.bili-share-to-friends-btn-primary {
-  color: #fff;
-  border-color: #00aeec;
-  background: #00aeec;
-}
-
-.bili-share-to-friends-btn-primary:hover {
-  color: #fff;
-  border-color: #40c5f1;
-  background: #40c5f1;
-}
-
-.bili-share-to-friends-btn:disabled {
-  border-color: #e3e5e7;
-  background: #f6f7f8;
-  color: #c9ccd0;
-  cursor: not-allowed;
 }
